chore(ml): util系の関数の追加
